New PDFStreamProcessor base of PDFStreamEngine adds simple stream processor
https://issues.apache.org/jira/browse/PDFBOX-1256

PDFStreamEngine mixes the functionality of an event-driven stream processor
with a content-stream-aware stateful engine. That makes PDFStreamProcessor
clumsy to use with code that doesn't need or want its awareness of the graphics
state stack, etc.

This patch separates the two roles, moving the simple event-driven content
stream functionality into the new base class PDFStreamEngine and making
PDFStreamProcessor a specialisation of it.

The new PDFStreamProcessor may be used for any task where a PDF stream must be
scanned, making no assumptions as to the validity of the operators, etc. It
also adds a fallback/default handler that's invoked if no handler was explicitly
set for a particular operator.

An interface for operator callbacks, PDFStreamProcessor.OperatorHandler,
is added, and the existing OperatorHandler abstract class is adjusted to
implement it. This is necessary because OperatorHandler is directly
tied to the PDFStreamEngine type; it also allows implementers to choose
not to have a ref to the stream processor if they don't need one, to use
an inner class, etc instead of using the set/getContext methods.