Major rework of Overlay.java to use PDFStreamProcessor for stream rewriting
Overlay.java has to rename objects in resource dictionaries, and their
references from the content stream(s), when the names in the overlay conflict
with names that already appear in the target document.

It used to do this with a hand-rolled primitive content stream parser/rewriter
that basically did binary substitution of names in the stream after some
correction for escaped and in-string contents. To generate new names it'd
append "overlay" and hope that would make the new name unique.

This rewrite changes both the name generation strategy and the way streams are
rewritten.  It makes Overlay.java pass streams through a PDFStreamProcessor
that copies each operation (operator + arguments) found from the input stream
to the output stream. It matches certain operators that refer to names in the
resources dictionary and rewrites those names if they are known to clash.
Clashes are determined by building a set of all names used in the target
document's resources dictionaries before beginning the rewriting process. When
new names are generated for uniqueness, "On" is appended, where "n" is a
counter incremented until a unique non-conflicting name is found.

Currently Overlay.java can still generate names longer than the legal maximum,
though it's much less likely to do so than the old version. Further changes are
required to detect overlength names and rewrite them differently.